pyproject.toml
[project]
name = "voxcortex"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "pydantic>=2.6",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "google-cloud-pubsub>=2.20",
]

[tool.uvicorn]
factory = false
